refactor(ajv): single-source AJV construction in @lobu/core (Task C, scoped)
Task C's "drop the CLI's local Ajv, validate against TypeBox contracts" rests on
a misdiagnosis: the CLI's AJV in `apply/desired-state.ts` validates operator
config against arbitrary RUNTIME connector JSON Schemas (a connector-author-
declared `optionsSchema`), which TypeBox can't replace — TypeBox validates
compile-time-known schemas, not dynamic JSON Schema documents.

The real duplication was two drifting AJV singletons: the server's
`ajv-singleton.ts` (hardened `allErrors:false` against a CWE-400 error-
amplification DoS) and the CLI's copy (`allErrors:true`, no such hardening).

Hoist a `createAjv(options)` factory + `formatAjvError` into `@lobu/core/ajv`
(new subpath export so the browser bundle never pulls AJV). Each caller keeps
its intentional options — the server fails fast + coerces on untrusted input;
the CLI reports every error over the operator's own local config — but the
construction + formatter are now single-sourced. Also migrated
`operations/input-validation.ts` to `createAjv` (identical config, zero
behavior change). `complete-window.ts` keeps its bespoke
`removeAdditional:true`/no-formats setup and is intentionally left.

Verify: make test-e2e-cli (66/0 — incl. `lobu apply` dry-run + real, the
connector-config validation path); server unit 566/0; core/cli/server tsc;
make pre-pr clean.
